add private python api for memcpy op
